third_party/WebKit/LayoutTests/inspector-enabled/console/console-uncaught-promise-no-inspector-expected.txt

CONSOLE ERROR: line 9: Unhandled promise rejection
CONSOLE ERROR: line 28: Unhandled promise rejection
CONSOLE ERROR: line 47: Unhandled promise rejection
Tests that uncaught promise rejection messages have line numbers when the inspector is closed and stack traces are not collected.

console-uncaught-promise-no-inspector.html:9 Unhandled promise rejection Promise
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: err1
console-uncaught-promise-no-inspector.html:28 Unhandled promise rejection Promise
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: Error: err2
console-uncaught-promise-no-inspector.html:47 Unhandled promise rejection Promise
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
